refactor(phase29): complete architecture deep cleanup

+++ lib/features/home/presentation/widgets/home_feed_view_utils.dart
@@ -38,7 +38,7 @@ double resolveHomeFeedCacheExtent(
     value *= tuning.reducedEffectsFactor;
   }
 
-  return value.clamp(tuning.minExtent, base).toDouble();
+  return value.clamp(tuning.minExtent, base);
 }
 
 void prefetchHomeFeedImages(
